Stabilize the TestAlertingRebalancerFailure test (#608)
TestAlertingRebalancerFailure was made stable.

This test was unstable because there were some race condition between two tests (cleanup of testParticipantUnavailable and start of the testTagSetIncorrect). Due to this race condition, the listfield and mapfield in the IdealState have been missing for the testTagSetIncorrect test. Hence, rebalancer does not run as expected.