Update listClusters() in ZkHelixAdmin (#895)
This PR updates getClusters() in ZkHelixAdmin in realm-aware mode. It is difficult to reason about getting all clusters - and this method, on single-realm mode, is also broken anyways because in the case users create nested clusters, this method will end up returning nothing. This update is backward-compatible for single-realm users.

Ideally, users should use Metadata Store Directory Service to get the list of all clusters.